refactor(structure) Move lib to src
- Rename lib/ directory to src/ for modern conventions
- Update all imports in tests, examples, and config
- Update package.json exports, main, types, and files
- Update tsconfig.json include paths
- Add doc/ to published files
