+++ examples/async-uncaught-exception.js
@@ -1,5 +1,5 @@
 import fs from 'node:fs'
-import errs from '../lib/errs.js'
+import errs from '../src/errs.js'
 
 process.on('uncaughtException', err => {
   console.log(errs.merge(err, { namespace: 'uncaughtException' }))
