+++ src/errs.d.ts
