+++ src/format.js
