+++ test/format.test.js
@@ -8,7 +8,7 @@
 
 import { strict as assert } from 'node:assert'
 import { describe, it } from 'node:test'
-import { format } from '../lib/format.js'
+import { format } from '../src/format.js'
 
 describe('format', () => {
   describe('terminal format', () => {
